requirements.txt
fastapi==0.116.1
langchain==0.3.27
langchain-openai==0.3.30
prisma==0.15.0
psycopg2-binary==2.9.10
python-dotenv==1.1.1
uvicorn==0.35.0
pypdf
langchain-community
faiss-cpu==1.7.4  # or whatever version you prefer